Empty HU write-off — decide the emptying eligibility at job-load time
