+++ backend/de.metas.manufacturing.rest-api/src/main/java/de/metas/manufacturing/job/service/ManufacturingJobLoaderAndSaverSupportingServices.java
@@ -6,6 +6,7 @@ import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Multimaps;
 import de.metas.handlingunits.HuId;
 import de.metas.handlingunits.IHandlingUnitsBL;
+import de.metas.handlingunits.generichumodel.HUType;
 import de.metas.handlingunits.model.I_M_HU;
 import de.metas.handlingunits.pporder.api.IHUPPOrderBL;
 import de.metas.handlingunits.pporder.api.issue_schedule.PPOrderIssueSchedule;
@@ -164,5 +165,41 @@ public class ManufacturingJobLoaderAndSaverSupportingServices
 	{
 		final ImmutableSet<HuId> huIds = sourceHUService.getSourceHUIds(ppOrderId);
 		return handlingUnitsBL.getLocatorIds(huIds);
+	}
+
+	/**
+	 * @return {@code true} if the HU's shape (not aggregate, a {@link HUType#TransportUnit} or
+	 * {@link HUType#VirtualPI} carrying at most one product storage) allows it to later be written off
+	 * via the "empty (auto. inventory)" reason. The client-config gate (whether emptying is offered at
+	 * all) is deliberately not applied here -- it is resolved once per request by the caller.
+	 */
+	public boolean isEmptyingEligible(@NonNull final HuId huId)
+	{
+		final I_M_HU hu = handlingUnitsBL.getById(huId);
+		return isEmptyingEligible(hu);
+	}
+
+	private boolean isEmptyingEligible(@NonNull final I_M_HU hu)
+	{
+		if (handlingUnitsBL.isAggregateHU(hu)) { return false; }          // orthogonal to unit type
+
+		final HUType huType = HUType.ofCodeOrNull(handlingUnitsBL.getHU_UnitType(hu));
+		if (huType == null) { return false; }                             // getHU_UnitType is @Nullable
+
+		switch (huType)
+		{
+			case TransportUnit:
+			case VirtualPI:
+				return isSingleProductStorage(hu);
+			case LoadLogistiqueUnit:
+				return false;
+			default:
+				return false;                                             // a future unit type is not silently accepted
+		}
+	}
+
+	private boolean isSingleProductStorage(@NonNull final I_M_HU hu)
+	{
+		return handlingUnitsBL.getStorageFactory().getProductStorages(hu).size() <= 1;
 	}
 }
